handle case when object is not created yet

+++ apps/search/tasks.py
@@ -14,18 +14,20 @@ def update_or_create_user_search_object_task(instance_pk):
 
 def _update_or_create_user_search_object_task(instance_pk):
     """Create the associated search object at people group's creation."""
-    user = ProjectUser.objects.get(pk=instance_pk)
-    search_objects = SearchObject.objects.filter(
-        type=SearchObject.SearchObjectType.USER, user=user
-    )
-    while search_objects.count() > 1:
-        search_objects.last().delete()
-    search_object, _ = SearchObject.objects.update_or_create(
-        user=user,
-        type=SearchObject.SearchObjectType.USER,
-        defaults={"last_update": timezone.localtime(timezone.now())},
-    )
-    search_object.save()
+    user = ProjectUser.objects.filter(pk=instance_pk)
+    if user.exists():
+        user = user.get()
+        search_objects = SearchObject.objects.filter(
+            type=SearchObject.SearchObjectType.USER, user=user
+        )
+        while search_objects.count() > 1:
+            search_objects.last().delete()
+        search_object, _ = SearchObject.objects.update_or_create(
+            user=user,
+            type=SearchObject.SearchObjectType.USER,
+            defaults={"last_update": timezone.localtime(timezone.now())},
+        )
+        search_object.save()
 
 
 @app.task(name="apps.search.tasks.update_or_create_project_search_object")
@@ -35,18 +37,20 @@ def update_or_create_project_search_object_task(instance_pk):
 
 def _update_or_create_project_search_object_task(instance_pk):
     """Create the associated search object at people group's creation."""
-    project = Project.objects.get(pk=instance_pk)
-    search_objects = SearchObject.objects.filter(
-        type=SearchObject.SearchObjectType.PROJECT, project=project
-    )
-    while search_objects.count() > 1:
-        search_objects.last().delete()
-    search_object, _ = SearchObject.objects.update_or_create(
-        project=project,
-        type=SearchObject.SearchObjectType.PROJECT,
-        defaults={"last_update": timezone.localtime(timezone.now())},
-    )
-    search_object.save()
+    project = Project.objects.filter(pk=instance_pk)
+    if project.exists():
+        project = project.get()
+        search_objects = SearchObject.objects.filter(
+            type=SearchObject.SearchObjectType.PROJECT, project=project
+        )
+        while search_objects.count() > 1:
+            search_objects.last().delete()
+        search_object, _ = SearchObject.objects.update_or_create(
+            project=project,
+            type=SearchObject.SearchObjectType.PROJECT,
+            defaults={"last_update": timezone.localtime(timezone.now())},
+        )
+        search_object.save()
 
 
 @app.task(name="apps.search.tasks.delete_project_search_object")
@@ -68,15 +72,17 @@ def update_or_create_people_group_search_object_task(instance_pk):
 
 def _update_or_create_people_group_search_object_task(instance_pk):
     """Create the associated search object at people group's creation."""
-    people_group = PeopleGroup.objects.get(pk=instance_pk)
-    search_objects = SearchObject.objects.filter(
-        type=SearchObject.SearchObjectType.PEOPLE_GROUP, people_group=people_group
-    )
-    while search_objects.count() > 1:
-        search_objects.last().delete()
-    search_object, _ = SearchObject.objects.update_or_create(
-        people_group=people_group,
-        type=SearchObject.SearchObjectType.PEOPLE_GROUP,
-        defaults={"last_update": timezone.localtime(timezone.now())},
-    )
-    search_object.save()
+    people_group = PeopleGroup.objects.filter(pk=instance_pk)
+    if people_group.exists():
+        people_group = people_group.get()
+        search_objects = SearchObject.objects.filter(
+            type=SearchObject.SearchObjectType.PEOPLE_GROUP, people_group=people_group
+        )
+        while search_objects.count() > 1:
+            search_objects.last().delete()
+        search_object, _ = SearchObject.objects.update_or_create(
+            people_group=people_group,
+            type=SearchObject.SearchObjectType.PEOPLE_GROUP,
+            defaults={"last_update": timezone.localtime(timezone.now())},
+        )
+        search_object.save()
